Make boot volume setup script useful on non-Mac platforms
Senses if running on macOS and:

1. Presents messages as notifications (as before)
2. Expands `eject__VolumeWithPrejudice` (a previously un-supplied
script) to be inline. In general, the "crud" macOS puts on volumes to
support Spotlight metadata does no harm to non-Mac systems but I prefer
to be tidy and remove it during the "eject" process.

If not running on macOS:

1. Presents messages via `echo`, which should work on Linux.
2. Bypasses Mac-specific cleanup. No attempt is made to eject the
volume on non-Mac systems. A placeholder is left as an exercise for
the reader.