Improved default handling of isc-dhcp-fix.sh
Previously, `isc-dhcp-fix.sh` was installed in `/usr/bin/` and a patch was applied to `/etc/rc.local` that included:

```
/usr/bin/isc-dhcp-fix.sh eth0 wlan0 &
```

In other words, there was an assumption that every Raspberry Pi had both Ethernet and WiFi interfaces, that both interfaces were active, and both obtained their IP addresses via DHCP.

This mod changes the patch applied to `/etc/rc.local` so that the default situation is:

```
```

Thus the starting position is `isc-dhcp-fix.sh` is installed but it is not activated. Even if the line is uncommented, it still amounts to a no-op because `isc-dhcp-fix.sh` exits if it is not provided with a list of interfaces.

The 02 script now checks both `eth0` `wlan0`. If and only if an interface is defined, active, and has obtained its IP address from DHCP is it included. The commented-out command is replaced with an active command that only includes the interfaces that have passed those tests.

There is an alternate scenario where a user has provided a host-specific patch. In that case, the "signature" of the default patch (`# /usr/bin/isc-dhcp-fix.sh &`) will not be present so there is no replacement.