disable IPv6 on GRUB; prerequisites; bump SQLite version
Adds support for GRUB-based boots. Default patch disables IPv6. The
need for this is really because of Network Manager which starts IPv6
services despite any settings in `sysctl.d`.

Improved handling of python3-ykman. The test now is "not buster" so it
is installed on both Bullseye and Bookworm. Previously, it was only
installed on Bullseye.

Bump SQLite version to 3420000.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>