reset exim4 paniclog if non-zero length
On a PiBuilder run on a Raspberry Pi:

1. The 02 script disables IPv6;
2. The 03 script installs the `at` package, which installs `exim4` as
a dependency.

The default configuration for `exim4` assumes IPv6 and complains about
the problem by writing messages to its "paniclog" at the rate of one
per reboot:

```
socket bind() to port 25 for address ::1 failed: Cannot assign requested address: daemon abandoned
```

In turn, the non-zero-length paniclog causes "you have mail" messages at
login time:

```
exim paniclog /var/log/exim4/paniclog on prx-deb.mull.lgosys.com has non-zero size, mail system might be broken. Up to 10 lines are quoted below.
```

PiBuilder deals with this situation *on the Raspberry Pi* by waiting
until `at` installs `exim4` and then patching the default configuration
to remove the assumption of IPv6. The patch can't be installed ahead of
time because the configuration files do not exist.

However, it is not just `at` which can install `exim4` as a dependency;
it may even come preinstalled. Thus there is a more general problem of
residual paniclog messages accumulating between when PiBuilder disables
IPv6 in the 02 script and when the configuration is patched in the
03 script. Those also result in "you have mail" messages at login time.

This change resets any non-zero-length paniclog that exists at the time
the patch is applied.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>