adds `set_hostname.sh` helper script
`set_hostname.sh` mimics (in part) the approach of `raspi-config` to
changing the hostname but augments it with a best-efforts discovery
of any local domain name which may have been learned from DHCP.
Taken together, this is closer to the result obtained from running
the Debian ISO installer. It results in `/etc/hosts` gaining a
fully-qualified domain name and that, in turn, means the `hostname -d`
and `hostname -f` commands work.

Change-of-hostname functionality in the 01 script delegated to
`set_hostname.sh`.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>