better to invoke `set_hostname.sh` with sudo
Using `sudo` on discrete commands within the `set_hostname.sh` script
seems to trigger the message:

```
sudo: unable to resolve host «name»: Name or service not known
```

This problem goes away if the discrete uses of `sudo` are removed in
favour of invoking the whole script with elevated privileges.

Also fixes a lurking bug on the `else` arm of the check for `systemd`
where `/etc/hostname` is written to via (unprivileged) `echo`.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>