explicitly disable IPv6 on physical interfaces (NetworkManager)
NetworkManager initialises physical interfaces to `ipv6.method=auto`.
This causes NetworkManager to try to assign an IPv6 address to each
physical interface. The hook script added by b0383cf (2024-04-09) keeps
IPv6 disabled system-wide so the attempt fails, and NetworkManager
immediately retries. As well as consuming CPU cycles and getting
nowhere, each attempt triggers a warning in the system journal.

This change causes the 02 script to iterate the available interfaces
and change any instances of `ipv6.method=auto` to `disabled`. The
changes are applied on the reboot at the end of the 02 script.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>