recent Bookworm release breaks /etc/dphys-swapfile patching
The default version of `/etc/dphys-swapfile` supplied with recent
versions of Raspberry Pi OS Bookworm conflicts with PiBuilder's patching
mechanism in a way that is not easily reconciled.

A new `try_edit()` function has been implemented via which a file of
`sed`editing commands can be processed to edit files in-situ. That is
now being used to edit `/etc/dphys-swapfile` in a more release-agnostic
manner.

If PiBuilder senses the presence of an old-style patch file for
`/etc/dphys-swapfile`, it displays a deprecation notice and forces
`VM_SWAP=default` which amounts to just leaving things alone.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>