implements Discord suggestions
adds new "automating your backups" section which includes
cross-references to IOTstackBackup configuration and crontab
instructions.

moves internal link target for words "Make a note of your `«runtag»`"
to the start of the numbered section containing those words. The
intention is that when the reader clicks the subsequent link to this
target, the screen will redraw to show the full context (of what a
`«runtag»` is and how to work it out) rather than just the reminder
to make a note of it which then requires the reader to scroll up to
reveal the context.

also fixes a GitHub cross-reference to adopt the new "tab=" URL.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>